refactor(rooch-da): modularize Indexer and TxPosition structs
Extracted Indexer and TxPosition into separate files and modules, improving code organization and reusability. This change reduces redundancy and streamlines dependencies within the `da` commands.
